Convert latin1 files to utf-8
The source files are parsed by $(make docs). Currently this is done
using python2 (actually 'python' which is still python2 on most
machines). When using python3 however (which has a stricter encoding
handling) it is easier if all input is in utf-8 encoding. So convert the
few remaining latin1 encoded files to utf-8. This only affects code
comments.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>