state: backend_bucket_circular: harmonize code with dt-utils
Comparing barebox' commit [1] with dt-utils' commit [2] it is obvious that
adjusting the return value has been forgotten in barebox - fix that.
(As the actual code that is changed is removed by the preprocesser when
compiling barebox this patch is only of cosmetic nature to keep both code bases
in sync).

[1] 9d6d91931afb ("state: Remove -EUCLEAN check from userspace tool")
[2] 791a2404116d ("state: Remove -EUCLEAN check from userspace tool")

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>